ISIS-1976: make ObjectAdapter's Oid final
