+++ core/plugins/jdo-datanucleus-5/src/main/java/org/apache/isis/core/runtime/persistence/adapter/PojoAdapter.java
@@ -19,7 +19,9 @@
 
 package org.apache.isis.core.runtime.persistence.adapter;
 
-import java.util.Objects;
+import org.datanucleus.enhancement.Persistable;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import org.apache.isis.core.commons.authentication.AuthenticationSession;
 import org.apache.isis.core.commons.exceptions.IsisException;
@@ -39,20 +41,18 @@ import org.apache.isis.core.metamodel.spec.ObjectSpecification;
 import org.apache.isis.core.metamodel.spec.Specification;
 import org.apache.isis.core.metamodel.spec.feature.OneToManyAssociation;
 import org.apache.isis.core.metamodel.specloader.SpecificationLoader;
+import org.apache.isis.core.runtime.system.persistence.PersistenceSession;
 import org.apache.isis.core.runtime.system.persistence.PersistenceSession5;
-import org.datanucleus.enhancement.Persistable;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class PojoAdapter extends InstanceAbstract implements ObjectAdapter {
 
     private final static Logger LOG = LoggerFactory.getLogger(PojoAdapter.class);
 
-    // -- Constructor, fields, finalizer
+    // -- Constructor, fields
 
     private final AuthenticationSession authenticationSession;
     private final SpecificationLoader specificationLoader;
-    private final PersistenceSession5 persistenceSession;
+    private final PersistenceSession persistenceSession;
 
     /**
      * can be {@link #replacePojo(Object) replace}d.
@@ -61,7 +61,7 @@ public class PojoAdapter extends InstanceAbstract implements ObjectAdapter {
     /**
      * can be {@link #replaceOid(Oid) replace}d.
      */
-    private Oid oid;
+    private final Oid oid;
 
     /**
      * only for standalone or parented collections.
@@ -73,7 +73,7 @@ public class PojoAdapter extends InstanceAbstract implements ObjectAdapter {
             final Oid oid,
             final AuthenticationSession authenticationSession,
             final SpecificationLoader specificationLoader,
-            final PersistenceSession5 persistenceSession) {
+            final PersistenceSession persistenceSession) {
 
         this.persistenceSession = persistenceSession;
         this.specificationLoader = specificationLoader;
@@ -129,11 +129,11 @@ public class PojoAdapter extends InstanceAbstract implements ObjectAdapter {
         return oid;
     }
 
-    @Override
-    public void replaceOid(Oid persistedOid) {
-        Objects.requireNonNull(oid); // values have no oid, so cannot be replaced
-        this.oid = persistedOid;
-    }
+//    @Override
+//    public void replaceOid(Oid persistedOid) {
+//        Objects.requireNonNull(oid); // values have no oid, so cannot be replaced
+//        this.oid = persistedOid;
+//    }
 
 
     // -- isParentedCollection, isValue
@@ -437,6 +437,9 @@ public class PojoAdapter extends InstanceAbstract implements ObjectAdapter {
         throw new UnsupportedOperationException();
     }
 
-
+    @Override
+    public ObjectAdapter withOid(RootOid newOid) {
+        return new PojoAdapter(pojo, newOid, authenticationSession, specificationLoader, persistenceSession);
+    }
 
 }
